# Conflicts:
#	data/LB_heater.js.gz
#	data/ajaxloader.js.gz
#	data/bg.jpg
#	data/bootstrap.min.css.gz
#	data/bootstrap.min.js.gz
#	data/dialog.js.gz
#	data/index.htm
#	data/jquery.min.js.gz
#	data/jquery.touchSwipe.min.js.gz
#	data/scripts.js.gz
#	data/style.css.gz

+++ buzzer.h
@@ -1,16 +1,18 @@
-// buzzer.h
-
+//   Проект центра открытого проектирования у Счастливчика https://LuckyCenter.ru
 #ifndef _BUZZER_h
 #define _BUZZER_h
 
 #if defined(ARDUINO) && ARDUINO >= 100
-	#include "arduino.h"
+	#include "Arduino.h"
 #else
 	#include "WProgram.h"
 #endif
 
-extern void initBuzzer(unsigned long duration);
+#include "setting.h"
+#include "pwm_out.h"
+#include "user_config.h"
+
+extern void initBuzzer(uint16_t duration);
 extern void deinitBuzzer();
 
 #endif
-
